+++ finagle-http/src/main/scala/com/twitter/finagle/http/filter/StatsFilter.scala
@@ -11,8 +11,8 @@ object StatsFilter {
 
   def module: Stackable[ServiceFactory[Request, Response]] =
     new Stack.Module1[param.Stats, ServiceFactory[Request, Response]] {
-      val role = StatsFilter.role
-      val description = StatsFilter.description
+      val role: Stack.Role = StatsFilter.role
+      val description: String = StatsFilter.description
 
       def make(
         statsParam: param.Stats,
@@ -44,8 +44,6 @@ class StatsFilter[REQUEST <: Request](stats: StatsReceiver)
   private[this] val statusReceiver = stats.scope("status")
   private[this] val timeReceiver = stats.scope("time")
   private[this] val responseSizeStat = stats.stat("response_size")
-  private[this] val requestStreamDurationMs = stats.stat("request_stream_duration_ms")
-  private[this] val responseStreamDurationMs = stats.stat("response_stream_duration_ms")
 
   private[this] val counterCache: String => Counter =
     Memoize(statusReceiver.counter(_))
@@ -56,15 +54,9 @@ class StatsFilter[REQUEST <: Request](stats: StatsReceiver)
   def apply(request: REQUEST, service: Service[REQUEST, Response]): Future[Response] = {
     val elapsed = Stopwatch.start()
 
-    if (request.isChunked) {
-      countRequestStreamDuration(request)
-    }
     val future = service(request)
     future respond {
       case Return(response) =>
-        if (response.isChunked) {
-          countResponseStreamDuration(response)
-        }
         count(elapsed(), response)
       case Throw(_) =>
         // Treat exceptions as empty 500 errors
@@ -85,17 +77,5 @@ class StatsFilter[REQUEST <: Request](stats: StatsReceiver)
     statCache(statusClass).add(duration.inMilliseconds)
 
     responseSizeStat.add(response.length)
-  }
-
-  private def countRequestStreamDuration(request: REQUEST): Unit = {
-    val streamingRequestElapsed = Stopwatch.start()
-    request.reader.onClose.respond(_ =>
-      requestStreamDurationMs.add(streamingRequestElapsed().inMilliseconds))
-  }
-
-  private def countResponseStreamDuration(response: Response): Unit = {
-    val streamingResponseElapsed = Stopwatch.start()
-    response.reader.onClose.respond(_ =>
-      responseStreamDurationMs.add(streamingResponseElapsed().inMilliseconds))
   }
 }